packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "limiq-sdk"
version = "0.1.0"
description = "Limiq.io Python SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.1",
  "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/limiq_sdk"]

[tool.pytest.ini_options]
pythonpath = ["src"]
